requirements.txt
google-generativeai>=0.7.2
python-dotenv>=1.0.1
mcp>=0.1.0
rich>=13.7.1
PyYAML>=6.0.1
